+++ src/main/java/cmdUI/UserFunctions.java
@@ -1,28 +1,39 @@
 package cmdUI;
 
-import task.taskEntities.Task;
 import printers.TaskListPrinter;
 import printers.TimelinePrinter;
-import task.taskEntities.EventTask;
+
+import task.taskController.TaskCreator;
+import task.taskUseCases.TaskTracker;
+import timeline.TimelineManager;
 import task.tasklistEntities.TaskList;
 import timeline.Timeline;
-import task.taskUseCases.TaskTracker;
-import dateAndTime.dateAndTimeAttributes.DateRange;
-import dateAndTime.dateAndTimeAttributes.TimeRange;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.Scanner;
 
 public class UserFunctions{
+    TaskList tasklist;
+    TaskTracker tracker;
+    Timeline timeline;
+    TimelineManager timelineManager;
 
-    public static void mainMenu(Timeline timeline, TaskTracker tracker) {
+    public UserFunctions() {
+        // TODO: Remove instances of TaskList and Timeline (Entities)
+        this.tasklist = new TaskList();
+        this.tracker = new TaskTracker(this.tasklist);
+        this.timeline = new Timeline();
+        this.timelineManager = new TimelineManager(this.timeline);
+    }
+
+    // Begin prompt which allows user to access features
+    public void mainMenu() {
         System.out.println("Welcome back!");
         System.out.println("What would you like to do?");
         String msg = "\nM to manage tasks" +
-                     "\nP to show information" +
-                     "\nX to exit";
+                "\nP to print information" +
+                "\nX to exit";
 
+        // Repeat message until user chooses to exit
         while (true) {
             System.out.println(msg);
 
@@ -30,10 +41,10 @@ public class UserFunctions{
             String key = reader.next();
 
             if (key.equalsIgnoreCase("M")) {
-                manageTask(timeline, tracker);
+                manageTask();
             }
             else if (key.equalsIgnoreCase("P")) {
-                printer(timeline,tracker);
+                printer();
             }
             else if (key.equalsIgnoreCase("X")) {
                 break;
@@ -41,73 +52,48 @@ public class UserFunctions{
         }
     }
 
-    // ask user what to do, and access the corresponding method in abstractClasses.Task tracker
-    public static void manageTask(Timeline timeline, TaskTracker tracker){
+    // Ask user what to do, and access the corresponding method in the TaskTracker
+    public void manageTask(){
         String msg = "\nA to create and add a task" +
-                     "\nENTER to exit";
+                "\nENTER to exit";
 
         System.out.println(msg);
 
         Scanner reader = new Scanner(System.in);
         String key = reader.nextLine();
 
+        // Call createTask
         if (key.equalsIgnoreCase("A")) {
-            taskCreation(tracker);
+            TaskCreator.createTask(this.tracker);
         }
     }
 
-    // print either storableEntities.Timeline, TaskList or Suggestion list
-    public static void printer(Timeline timeline, TaskTracker tracker){
+    // Print either Timeline, TaskList or Suggestion list
+    public void printer(){
         String msg = "\nM to show timeline" +
-                     "\nL to show to-do list" +
-                     "\nS to show suggestion list" +
-                     "\nENTER to exit";
-
+                "\nL to show to-do list" +
+                "\nS to show suggestion list" +
+                "\nENTER to exit";
 
         System.out.println(msg);
 
         Scanner reader = new Scanner(System.in);
         String key = reader.nextLine();
 
+        // Call print method in respective printer classes
         if (key.equalsIgnoreCase("M")) {
-            TimelinePrinter.print(timeline);
+            TimelinePrinter.print(this.timeline);
         }
         else if (key.equalsIgnoreCase("L")) {
-            TaskList tasklist = tracker.getTaskList();
-            TaskListPrinter.print(tasklist);
+            TaskListPrinter.print(this.tracker);
         }
         else if (key.equalsIgnoreCase("S")) {
             System.out.println("Not implemented yet");
         }
-
     }
 
-    // method about pomodoro
+    // TODO: Add method for Pomodoro
 
 
-
-    // Some helper methods
-    public static void taskCreation(TaskTracker tracker){
-        Scanner reader = new Scanner(System.in);
-
-        System.out.println("Please input the task name");
-        String name = reader.nextLine();
-        System.out.println("Please input the priority");
-        String priority = reader.nextLine();
-        System.out.println("Please input the task description");
-        String description = reader.nextLine();
-        System.out.println("Please input the start date of the task (YYYY-MM-DD)");
-        String startDate = reader.nextLine();
-        System.out.println("Please input the start time of the task (HH:mm)");
-        String startTime = reader.nextLine();
-        System.out.println("Please input the end date of the task (YYYY-MM-DD)");
-        String endDate = reader.nextLine();
-        System.out.println("Please input the end time of the task (HH:mm)");
-        String endTime = reader.nextLine();
-        DateRange dateRange = new DateRange(LocalDate.parse(startDate), LocalDate.parse(endDate));
-        TimeRange timeRange = new TimeRange(LocalTime.parse(startTime), LocalTime.parse(endTime));
-        Task task = new EventTask(name,priority,description, dateRange, timeRange);
-        tracker.addTask(task);
-    }
 
 }
